refactor(controller): split collection export module
