+++ src/egui_app/controller/collection_export/paths.rs
@@ -1,0 +1,48 @@
+use super::super::*;
+use crate::sample_sources::collections::CollectionMember;
+use std::path::{Path, PathBuf};
+
+pub(in crate::egui_app::controller) fn resolved_export_dir(
+    collection: &Collection,
+    global_root: Option<&Path>,
+) -> Option<PathBuf> {
+    if let Some(path) = collection.export_path.clone() {
+        Some(path)
+    } else {
+        global_root.map(|root| {
+            crate::sample_sources::config::normalize_path(
+                root.join(collection_folder_name(collection)).as_path(),
+            )
+        })
+    }
+}
+
+pub(in crate::egui_app::controller) fn export_dir_for(
+    collection: &Collection,
+    global_root: Option<&Path>,
+) -> Result<PathBuf, String> {
+    resolved_export_dir(collection, global_root).ok_or_else(|| "Set an export folder first".into())
+}
+
+pub(in crate::egui_app::controller) fn collection_folder_name(collection: &Collection) -> String {
+    collection.export_folder_name()
+}
+
+pub(in crate::egui_app::controller) fn delete_exported_file(
+    export_dir: Option<PathBuf>,
+    member: &CollectionMember,
+) {
+    let Some(dir) = export_dir else {
+        return;
+    };
+    let file_name = match member.relative_path.file_name() {
+        Some(name) => name,
+        None => return,
+    };
+    let target = dir.join(file_name);
+    let _ = std::fs::remove_file(target);
+}
+
+pub(in crate::egui_app::controller) fn collection_folder_name_from_str(name: &str) -> String {
+    crate::sample_sources::collections::collection_folder_name_from_str(name)
+}
